refactor: 패키지명 네이밍 규칙 통일 - 카멜 규칙 적용 - utils 폴더 -> validation으로 패키지명 수정
