+++ src/main/java/lotto/view/LottoGameView.java
@@ -1,4 +1,4 @@
-package lotto.View;
+package lotto.view;
 
 public class LottoGameView {
 }
