+++ src/main/java/lotto/view/PurchaseView.java
@@ -1,4 +1,4 @@
-package lotto.View;
+package lotto.view;
 
 public class PurchaseView {
 }
